Changed .sort(Comparator) to be a default method in Collection.
