+++ src/main/java/co/tsyba/core/collections/Set.java
@@ -1,7 +1,5 @@
 package co.tsyba.core.collections;
 
-import java.util.Arrays;
-import java.util.Comparator;
 import java.util.HashSet;
 import java.util.function.Function;
 import java.util.function.Predicate;
@@ -205,21 +203,6 @@ public class Set<T> extends RobinHoodHashStore<T> implements Collection<T> {
 
 		product.removeExcessCapacity();
 		return product;
-	}
-
-	@Override
-	public List<T> sort(Comparator<T> comparator) {
-		@SuppressWarnings("unchecked")
-		final var items = (T[]) new Object[entryCount];
-		var index = 0;
-
-		for (var item : this) {
-			items[index] = item;
-			++index;
-		}
-
-		Arrays.sort(items, comparator);
-		return new List<>(items);
 	}
 
 	@Override
